requirements.txt
Pillow
requests
pytest